package/ai/requirements-gpu.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
insightface>=0.7.3
onnxruntime-gpu>=1.17.0
numpy>=1.26.0
pillow
opencv-python-headless>=4.9.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
modelscope
# Torch (适配GPU版本)
# 用户应根据其硬件安装相应版本
# GPU版本（推荐CUDA 11.8）：
torch>=2.0.0+cu118
torchvision>=0.15.0+cu118

# PaddleOCR
paddlepaddle-gpu==3.2.0
paddleocr==3.3.2
